+++ Modules/SWMB/Custom-VarDefault.psm1
@@ -72,4 +72,9 @@ $Global:SWMB_Custom = @{
 	# Global cadence in days - Every day = 1 - Every week = 7 - Every month = 30 - During low free disk space = 0
 	StorageSenseCadence      =  30 # every month
 	StorageSenseTrashCleanup = 120 # delete files in recycle bin that are more than 120 days
+
+	# AutoLogon
+	# Use by tweak : EnableAutoLogon
+	# Inactive if empty
+	AutoLogon_UserName = ""
 }