+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Version 3.17 (in progress)
 
 New presets/rules:
+* 2024/04/10 - `DisableAutoLogon`/`EnableAutoLogon`/`ViewAutoLogon` -> Disable or Enable AutoLogon (`$Global:SWMB_Custom` key `AutoLogon_UserName`)
 * 2024/04/09 - `DisableUpgradesOnUnsupportedHard`/`EnableUpgradesOnUnsupportedHard`/`ViewUpgradesOnUnsupportedHard` -> Force a registry key so that hardware not supported for accepting Windows 11 upgrades can continue to benefit from updates (use at your own risk).
 
 ## Version 3.16